Fix UnifyingResolution

+++ src/main/scala/at/logic/skeptik/parser/ProofParserSPASS.scala
@@ -12,6 +12,8 @@ import at.logic.skeptik.expression.formula._
 import at.logic.skeptik.expression._
 import at.logic.skeptik.judgment.immutable.{ SeqSequent => Sequent }
 import at.logic.skeptik.proof.sequent.resolution._
+import at.logic.skeptik.expression.substitution.immutable.Substitution
+
 
 object ProofParserSPASS extends ProofParser[Node] with SPASSParsers
 
@@ -62,56 +64,67 @@ trait SPASSParsers
       def firstNode = firstRef.first
       def secondNode = secondRef.first
 
-
       def firstClause = exprMap.getOrElse(firstRef.first + "." + firstRef.second, throw new Exception("Error!"))
       def secondClause = exprMap.getOrElse(secondRef.first + "." + secondRef.second, throw new Exception("Error!"))
-      
+
       //val ax = UnifyingResolution(proofMap.getOrElse(firstNode, throw new Exception("Error!")), proofMap.getOrElse(secondNode, throw new Exception("Error!")),
       //     exprMap.getOrElse(firstRef.first + "." + firstRef.second, throw new Exception("Error!")),exprMap.getOrElse(secondRef.first + "." + secondRef.second, throw new Exception("Error!")))(vars)
-    
+
       val firstPremise = proofMap.getOrElse(firstNode, throw new Exception("Error!"))
       val secondPremise = proofMap.getOrElse(secondNode, throw new Exception("Error!"))
-      
+
       println("First premise: " + firstPremise)
       println("First clause: " + firstClause)
       println("Second premise: " + secondPremise)
       println("Second clause: " + secondClause)
-      
+
       val ax = UnifyingResolution(firstPremise, secondPremise)(vars)
 
       //TODO: do we actually have to reverse this naming?
-      //TODO: check if this is null
-      //TODO: make this support multiple replacements
-      //TODO: replace ax with sFinal
-      val rev = ax.replacementInverse
       
-      val newAnt = for (a <- ax.conclusion.ant) yield rev(a)
-      
-      val sA = addAntecedentsSeq(newAnt)
+      if (ax.replacementInverse.length > 0) {
+//        val rev = ax.replacementInverse(0)
 
-      val sS = addSuccedentsSeq(ax.conclusion.suc)
+        def performReplacements(ant: Seq[E], replacements: List[Substitution]): Seq[E] ={
+          if (replacements.length > 0){
+            val newAnt = for (a <- ax.conclusion.ant) yield replacements.head(a)
+            performReplacements(newAnt, replacements.tail)
+          } else {
+            ant
+          }
+        }
+        
+        val newAnt = performReplacements(ax.conclusion.ant, ax.replacementInverse)// for (a <- ax.conclusion.ant) yield rev(a)
 
-      val sFinal = sA union sS
+        val sA = addAntecedentsSeq(newAnt)
 
-//      val ay = new Axiom(sFinal)
-//      proofMap += (ln -> ax)
-      
-      val parsedAnte = addAntecedents(seq._1)
+        val sS = addSuccedentsSeq(ax.conclusion.suc)
 
-      val parsedSucc = addSuccedents(seq._2)
+        val sFinal = sA union sS
 
-      val parsedFinal = parsedAnte union parsedSucc
+        //      val ay = new Axiom(sFinal)
+        //      proofMap += (ln -> ax)
 
-      val ay = new Axiom(parsedFinal)
-//      proofMap += (ln -> ax)
-//      count = ln
-      
-      println("Parsed: " + ln + ":" + ay)
-      println("Computed: " + ln + ":" + ax)
-      println("Computed B: " + ln + ":" + sFinal) //this one "matches" parsed
-      
-      proofMap += (ln -> ax)
-      ax
+        val parsedAnte = addAntecedents(seq._1)
+
+        val parsedSucc = addSuccedents(seq._2)
+
+        val parsedFinal = parsedAnte union parsedSucc
+
+        val ay = new Axiom(parsedFinal)
+        //      proofMap += (ln -> ax)
+        //      count = ln
+
+        println("Parsed: " + ln + ":" + ay)
+        println("Computed: " + ln + ":" + ax)
+        println("Computed B: " + ln + ":" + sFinal) //this one "matches" parsed
+        proofMap += (ln -> ay)
+        ay
+
+      } else {
+        proofMap += (ln -> ax)
+        ax
+      }
     }
 
     //For now, treat the other inference rules as new axioms
@@ -124,24 +137,23 @@ trait SPASSParsers
       val sFinal = sA union sS
 
       val ax = new Axiom(sFinal)
-      proofMap += (ln -> ax)      
+      proofMap += (ln -> ax)
       ax
     }
   }
 
   def sequent: Parser[(List[E], List[E])] = antecedent ~ "->" ~ succedent ~ "." ^^ {
     case ~(~(~(a, _), s), _) => {
-      
-      
+
       //This function maintains a map of the form ((proof line, clause position) -> clause).       
       def addToExprMap(lineNumber: Int, startPos: Int, exps: List[E]): Int = {
         if (exps.length > 1) {
           exprMap.getOrElseUpdate(lineNumber + "." + startPos, exps.head)
-//          println(lineNumber + "." + startPos + " == " + exps.head)
+          //          println(lineNumber + "." + startPos + " == " + exps.head)
           addToExprMap(lineNumber, startPos + 1, exps.tail)
         } else if (exps.length == 1) {
           exprMap.getOrElseUpdate(lineNumber + "." + startPos, exps.head)
-//          println(lineNumber + "." + startPos + " == " + exps.head)
+          //          println(lineNumber + "." + startPos + " == " + exps.head)
           startPos + 1
         } else {
           startPos
@@ -149,9 +161,7 @@ trait SPASSParsers
       }
 
       addToExprMap(count, addToExprMap(count, 0, a), s)
-	  
 
-      
       if (count % 500 == 0) { println(count + " lines parsed") }
       (a, s)
     }
@@ -290,7 +300,7 @@ trait SPASSParsers
       s0
     }
   }
-  
+
   def addAntecedentsSeq(antes: Seq[E]): Sequent = {
     if (antes.length > 1) {
       val s1 = antes.head +: addAntecedentsSeq(antes.tail)
@@ -317,8 +327,8 @@ trait SPASSParsers
       val s0 = Sequent()()
       s0
     }
-  }  
-  
+  }
+
 }
 
 class Ref(f: Int, s: Int) {
